+++ paper/src/main/java/com/badbones69/crazycrates/api/PrizeManager.java
@@ -62,7 +62,7 @@ public class PrizeManager {
             return false;
         }
 
-        final int wins = userManager.getCrateRespin(player.getUniqueId(), crate.getFileName(), crate.isCyclePersistRestart());
+        final int wins = userManager.getCrateRespin(player.getUniqueId(), crate.getFileName());
 
         int cap = getCap(crate, player);
 
@@ -283,7 +283,7 @@ public class PrizeManager {
     public static @Nullable Tier getTier(@NotNull final Crate crate) {
         if (crate.getTiers().isEmpty()) return null;
 
-        final Random random = MiscUtils.useOtherRandom() ? ThreadLocalRandom.current() : new Random();
+        final Random random = MiscUtils.getRandom();
 
         double weight = 0.0;