+++ paper/src/main/java/com/badbones69/crazycrates/api/builders/types/features/CrateSpinMenu.java
@@ -7,8 +7,8 @@ import com.ryderbelserion.crazycrates.common.enums.Messages;
 import com.ryderbelserion.vital.paper.api.builders.gui.interfaces.Gui;
 import com.ryderbelserion.vital.paper.api.builders.gui.interfaces.GuiFiller;
 import com.ryderbelserion.vital.paper.api.builders.gui.interfaces.GuiItem;
+import com.ryderbelserion.vital.paper.util.scheduler.FoliaRunnable;
 import org.bukkit.entity.Player;
-
 import java.util.HashMap;
 import java.util.List;
 import java.util.UUID;
@@ -57,13 +57,18 @@ public class CrateSpinMenu extends StaticInventoryBuilder {
         this.gui.setOpenGuiAction(action -> this.userManager.addRespinPrize(uuid, fileName, this.settings.getPrize().getSectionName()));
 
         this.gui.setCloseGuiAction(action -> {
-            if (this.userManager.hasRespinPrize(uuid, fileName)) {
-                Messages.crate_prize_respin_not_claimed.sendMessage(player, new HashMap<>() {{
-                    put("{crate_pretty}", crate.getCrateName());
-                    put("{crate}", fileName);
-                    put("{prize}", userManager.getRespinPrize(uuid, fileName));
-                }});
-            }
+            new FoliaRunnable(this.plugin.getPlugin().getServer().getGlobalRegionScheduler()) {
+                @Override
+                public void run() {
+                    if (userManager.hasRespinPrize(uuid, fileName)) { // if they have a respin prize, add it.
+                        Messages.crate_prize_respin_not_claimed.sendMessage(player, new HashMap<>() {{
+                            put("{crate_pretty}", crate.getCrateName());
+                            put("{crate}", fileName);
+                            put("{prize}", userManager.getRespinPrize(uuid, fileName));
+                        }});
+                    }
+                }
+            }.runDelayed(this.plugin.getPlugin(), 20);
 
             this.crateManager.removePlayerFromOpeningList(this.player);
             this.crateManager.removeCrateInUse(this.player);