+++ paper/src/main/java/com/badbones69/crazycrates/commands/crates/types/admin/crates/CommandClaim.java
@@ -35,24 +35,33 @@ public class CommandClaim extends BaseCommand {
         }
 
         final UUID uuid = player.getUniqueId();
+        final String fileName = crate.getFileName();
 
-        if (!this.userManager.hasRespinPrize(uuid, crateName)) {
+        if (!this.userManager.hasRespinPrize(uuid, fileName)) {
             Messages.crate_prize_respins_empty.sendMessage(player, new HashMap<>() {{
                 put("{crate_pretty}", crate.getCrateName());
-                put("{crate}", crate.getFileName());
+                put("{crate}", fileName);
             }});
+
+            if (!crate.isCyclePersistRestart()) {
+                this.userManager.removeRespinCrate(uuid, fileName, this.userManager.getCrateRespin(uuid, fileName));
+            }
 
             return;
         }
 
-        final String prizeName = this.userManager.getRespinPrize(uuid, crateName);
+        final String prizeName = this.userManager.getRespinPrize(uuid, fileName);
 
         final Prize prize = crate.getPrize(prizeName);
 
         if (prize == null) {
             Messages.prize_not_found.sendMessage(player, "{prize}", prizeName);
 
-            this.userManager.removeRespinPrize(uuid, crateName);
+            if (!crate.isCyclePersistRestart()) {
+                this.userManager.removeRespinCrate(uuid, fileName, this.userManager.getCrateRespin(uuid, fileName));
+            }
+
+            this.userManager.removeRespinPrize(uuid, fileName);
 
             return;
         }
@@ -61,10 +70,14 @@ public class CommandClaim extends BaseCommand {
 
         Messages.crate_prize_respins_redeemed.sendMessage(player, new HashMap<>() {{
             put("{crate_pretty}", crate.getCrateName());
-            put("{crate}", crate.getFileName());
+            put("{crate}", fileName);
             put("{prize}", prize.getPrizeName());
         }});
 
-        this.userManager.removeRespinPrize(uuid, crateName);
+        if (!crate.isCyclePersistRestart()) {
+            this.userManager.removeRespinCrate(uuid, fileName, this.userManager.getCrateRespin(uuid, fileName));
+        }
+
+        this.userManager.removeRespinPrize(uuid, fileName);
     }
 }