+++ paper/src/main/java/com/badbones69/crazycrates/managers/BukkitUserManager.java
@@ -646,23 +646,13 @@ public class BukkitUserManager extends UserManager {
         this.data.save();
     }
 
-    private final Map<UUID, Map<String, Integer>> crates = new HashMap<>();
-
-    public int getCrateRespin(@NotNull final UUID uuid, @NotNull final String crateName, final boolean write) {
+    public int getCrateRespin(@NotNull final UUID uuid, @NotNull final String crateName) {
         final Crate crate = isCrateInvalid(crateName);
 
         if (crate == null) {
             if (MiscUtils.isLogging()) this.plugin.getLogger().warn("Crate {} doesn't exist.", crateName);
 
             return 0;
-        }
-
-        if (!write) {
-            if (!this.crates.containsKey(uuid)) {
-                return 0;
-            }
-
-            return this.crates.get(uuid).getOrDefault(crate.getFileName(), 0);
         }
 
         return this.data.getConfiguration().getInt("Players." + uuid + ".respins." + crateName + ".amount", 0);
@@ -740,17 +730,11 @@ public class BukkitUserManager extends UserManager {
         return configuration.getString("Players." + uuid + ".respins." + fileName + ".prize", "");
     }
 
-    public void removeRespinCrate(@NotNull final UUID uuid, @NotNull final String crateName, final int amount, final boolean write) {
+    public void removeRespinCrate(@NotNull final UUID uuid, @NotNull final String crateName, final int amount) {
         final Crate crate = isCrateInvalid(crateName);
 
         if (crate == null) {
             if (MiscUtils.isLogging()) this.plugin.getLogger().warn("Crate {} doesn't exist.", crateName);
-
-            return;
-        }
-
-        if (!write) {
-            this.crates.remove(uuid);
 
             return;
         }
@@ -772,7 +756,7 @@ public class BukkitUserManager extends UserManager {
         }
     }
 
-    public void addRespinCrate(@NotNull final UUID uuid, @NotNull final String crateName, final int amount, final boolean write) {
+    public void addRespinCrate(@NotNull final UUID uuid, @NotNull final String crateName, final int amount) {
         final Crate crate = isCrateInvalid(crateName);
 
         if (crate == null) {
@@ -782,24 +766,6 @@ public class BukkitUserManager extends UserManager {
         }
 
         final String fileName = crate.getFileName();
-
-        if (!write) {
-            if (this.crates.containsKey(uuid)) {
-                final Map<String, Integer> respins = this.crates.get(uuid);
-
-                respins.put(fileName, respins.get(fileName) + amount);
-
-                crates.put(uuid, respins);
-
-                return;
-            }
-
-            crates.put(uuid, new HashMap<>() {{
-                put(fileName, amount);
-            }});
-
-            return;
-        }
 
         final YamlConfiguration configuration = this.data.getConfiguration();