+++ paper/src/main/java/com/badbones69/crazycrates/tasks/crates/types/QuickCrate.java
@@ -39,7 +39,7 @@ public class QuickCrate extends CrateBuilder {
         super(crate, player, location);
     }
 
-    private final PluginManager server = this.plugin.getServer().getPluginManager();
+    private final PluginManager server = this.plugin.getPlugin().getServer().getPluginManager();
 
     private final Player player = getPlayer();
     private final Location location = getLocation();
@@ -144,8 +144,9 @@ public class QuickCrate extends CrateBuilder {
         } else {
             this.userManager.removeRespinPrize(this.uuid, fileName);
 
-            // remove from the cache
-            this.userManager.removeRespinCrate(this.uuid, fileName, 0, false);
+            if (!crate.isCyclePersistRestart()) {
+                userManager.removeRespinCrate(uuid, fileName, userManager.getCrateRespin(uuid, fileName));
+            }
         }
 
         PrizeManager.givePrize(this.player, this.crate, prize);