+++ paper/src/main/java/com/badbones69/crazycrates/tasks/crates/types/RouletteCrate.java
@@ -116,8 +116,9 @@ public class RouletteCrate extends CrateBuilder {
                             } else {
                                 userManager.removeRespinPrize(uuid, fileName);
 
-                                // remove from the cache
-                                userManager.removeRespinCrate(uuid, fileName, 0, false);
+                                if (!crate.isCyclePersistRestart()) {
+                                    userManager.removeRespinCrate(uuid, fileName, userManager.getCrateRespin(uuid, fileName));
+                                }
                             }
 
                             PrizeManager.givePrize(player, crate, prize);